Initial version 2.0 commit. Moved server into server directory.
