+++ plugins/eliza.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var utils = require('../utils.js');
+var utils = require('../server/utils.js');
 var elizabot = require('./elizabot.js');
 
 var commands = {
